Adds helper class to - assist in transforming back and forth between http/s and ws/s. This is useful during negotiate process - easily get a valid negotiate url irrespective of what is already in the path, query string or paramaters.
Updates HubConnectionBuilder try and 'negotiate' even for non-auth connections, this allows for unauthenticated Azure SignalR Service usage. If negotiation with Azure SignalR Service is successful, uses AuthConnectionHub instead of BaseConnectionHub

Updates AuthHubConnection to use helper class which appends 'negotiate' to the end of the path instead of the url in order to support the Azure SignalR Service which uses query parameters in its negotiate endpoint.